// backend/src/generated/type-graphql/resolvers/inputs/MovieListScalarWhereInput.ts
import * as TypeGraphQL from "type-graphql";
import * as GraphQLScalars from "graphql-scalars";
import { Prisma } from "@prisma/client";
import { DecimalJSScalar } from "../../scalars";
import { DateTimeFilter } from "../inputs/DateTimeFilter";
import { StringFilter } from "../inputs/StringFilter";
import { StringNullableListFilter } from "../inputs/StringNullableListFilter";

@TypeGraphQL.InputType("MovieListScalarWhereInput", {})
export class MovieListScalarWhereInput {
  @TypeGraphQL.Field(_type => [MovieListScalarWhereInput], {
    nullable: true
  })
  AND?: MovieListScalarWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [MovieListScalarWhereInput], {
    nullable: true
  })
  OR?: MovieListScalarWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [MovieListScalarWhereInput], {
    nullable: true
  })
  NOT?: MovieListScalarWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  id?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  createdAt?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  updatedAt?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  name?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  description?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableListFilter, {
    nullable: true
  })
  tags?: StringNullableListFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  userId?: StringFilter | undefined;
}
